fix(#730): break categories-read/models-read circular dependency
getModelWithCategoryMap and the count/tree joins (listCategoriesWithCounts,
getCategoryModelKitCounts, listCategoryTree) need BOTH domains' reads, so
keeping them in either categories-read.ts or models-read.ts made the two
import each other. Moved to a new model-category-join.ts, the one place
allowed to depend on both. Updates the two consumers (server/line-items.ts,
server/csv.ts) to import from the new module.

POLICY.md R-3.5.
