+++ src/server/csv.ts
@@ -6,7 +6,8 @@ import { prisma } from "@/lib/prisma";
 import { getConvexClient, toConvexDoc } from "@/lib/convex-client";
 import { api } from "../../convex/_generated/api";
 import { attachSupplier, getSuppliersByOrg } from "@/lib/suppliers-read";
-import { getModelWithCategoryMap, getModelsByOrg, type ConvexModel } from "@/lib/models-read";
+import { getModelsByOrg, type ConvexModel } from "@/lib/models-read";
+import { getModelWithCategoryMap } from "@/lib/model-category-join";
 import { getAssetsByOrg, getBulkAssetsByOrg, type ConvexAsset } from "@/lib/assets-read";
 import { getCategoryMap, getCategoriesByOrg } from "@/lib/categories-read";
 import { getLocationMap, getLocationsByOrg } from "@/lib/locations-read";
